+++ service/user.go
@@ -7,7 +7,6 @@ import (
 	"github.com/omnilaboratory/obd/bean/enum"
 	"github.com/omnilaboratory/obd/dao"
 	"github.com/omnilaboratory/obd/tool"
-	trackerBean "github.com/omnilaboratory/obd/tracker/bean"
 	"github.com/tyler-smith/go-bip39"
 	"log"
 	"time"
@@ -97,16 +96,4 @@ func (service *UserManager) UserLogout(user *bean.User) error {
 	}
 	noticeTrackerUserLogout(node)
 	return user.Db.Close()
-}
-
-func noticeTrackerUserLogin(user dao.User) {
-	loginRequest := trackerBean.ObdNodeUserLoginRequest{}
-	loginRequest.UserId = user.PeerId
-	sendMsgToTracker(enum.MsgType_Tracker_UserLogin_304, loginRequest)
-}
-
-func noticeTrackerUserLogout(user dao.User) {
-	loginRequest := trackerBean.ObdNodeUserLoginRequest{}
-	loginRequest.UserId = user.PeerId
-	sendMsgToTracker(enum.MsgType_Tracker_UserLogout_305, loginRequest)
 }
